+++ webui/src/components/BootstrapPanel.js
@@ -1,3 +1,4 @@
+// @flow
 import React from 'react';
 import { connect } from 'react-redux';
 import { css, cx } from 'emotion'
@@ -5,6 +6,7 @@ import { setVisibleBootstrapVshardPanel } from '../store/actions/clusterPage.act
 import { isBootstrapped, isRouterPresent, isStoragePresent } from '../store/selectors/clusterPage';
 import { IconCancel, IconOk, PageCard, Text } from '@tarantool.io/ui-kit';
 import type { State } from 'src/store/rootReducer';
+import { selectVshardRolesNames } from 'src/store/selectors/clusterPage';
 
 const styles = {
   iconMargin: css`
@@ -17,47 +19,58 @@ const styles = {
   `
 };
 
-class BootstrapPanel extends React.Component {
-  render() {
-    const {
-      bootstrapPanelVisible,
-      isBootstrapped,
-      requestingBootstrapVshard,
-      routerPresent,
-      storagePresent,
-      setVisibleBootstrapVshardPanel
-    } = this.props;
+type Props = {
+  bootstrapPanelVisible: bool,
+  isBootstrapped: bool,
+  requestingBootstrapVshard: bool,
+  routerPresent: bool,
+  storagePresent: bool,
+  setVisibleBootstrapVshardPanel: (v: bool) => void,
+  storageRoleName: string,
+  routerRoleName: string
+};
 
-    if (!bootstrapPanelVisible || requestingBootstrapVshard || isBootstrapped)
-      return null;
+const BootstrapPanel = (
+  {
+    bootstrapPanelVisible,
+    isBootstrapped,
+    requestingBootstrapVshard,
+    routerPresent,
+    storagePresent,
+    setVisibleBootstrapVshardPanel,
+    storageRoleName,
+    routerRoleName
+  }: Props
+) => {
+  if (!bootstrapPanelVisible || requestingBootstrapVshard || isBootstrapped)
+    return null;
 
-    return (
-      <PageCard
-        className='meta-test__BootstrapPanel'
-        title="Bootstrap vshard"
-        onClose={() => setVisibleBootstrapVshardPanel(false)}
-        showCorner
-      >
-        <Text className={styles.row} variant='h4'>
-          After you complete editing the topology, you need to bootstrap vshard to render storages operable.
-        </Text>
-        <Text className={styles.row}>
-          {routerPresent
-            ? <IconOk className={cx(styles.iconMargin, 'meta-test__BootstrapPanel__vshard-router_enabled')} />
-            : <IconCancel className={cx(styles.iconMargin, 'meta-test__BootstrapPanel__vshard-router_disabled')} />}
-          One role vshard-router enabled
-        </Text>
-        <Text className={styles.row}>
-          {storagePresent
-            ? <IconOk className={cx(styles.iconMargin, 'meta-test__BootstrapPanel__vshard-storage_enabled')} />
-            : <IconCancel className={cx(styles.iconMargin, 'meta-test__BootstrapPanel__vshard-storage_disabled')} />}
-          One role vshard-storage enabled
-        </Text>
-        <Text className={styles.row}>Afterwards, any change in topology will trigger data rebalancing</Text>
-      </PageCard>
-    );
-  }
-}
+  return (
+    <PageCard
+      className='meta-test__BootstrapPanel'
+      title="Bootstrap vshard"
+      onClose={() => setVisibleBootstrapVshardPanel(false)}
+      showCorner
+    >
+      <Text className={styles.row} variant='h4'>
+        After you complete editing the topology, you need to bootstrap vshard to render storages operable.
+      </Text>
+      <Text className={styles.row}>
+        {routerPresent
+          ? <IconOk className={cx(styles.iconMargin, 'meta-test__BootstrapPanel__vshard-router_enabled')} />
+          : <IconCancel className={cx(styles.iconMargin, 'meta-test__BootstrapPanel__vshard-router_disabled')} />}
+        {`One role ${routerRoleName} enabled`}
+      </Text>
+      <Text className={styles.row}>
+        {storagePresent
+          ? <IconOk className={cx(styles.iconMargin, 'meta-test__BootstrapPanel__vshard-storage_enabled')} />
+          : <IconCancel className={cx(styles.iconMargin, 'meta-test__BootstrapPanel__vshard-storage_disabled')} />}
+        {`One role ${storageRoleName} enabled`}
+      </Text>
+      <Text className={styles.row}>Afterwards, any change in topology will trigger data rebalancing</Text>
+    </PageCard>
+  );
+};
 
 const mapStateToProps = (state: State) => {
   const {
@@ -66,13 +79,16 @@ const mapStateToProps = (state: State) => {
       bootstrapPanelVisible
     }
   } = state;
+  const rolesNames = selectVshardRolesNames(state);
 
   return {
     bootstrapPanelVisible,
     isBootstrapped: isBootstrapped(state),
     requestingBootstrapVshard,
     routerPresent: isRouterPresent(state),
-    storagePresent: isStoragePresent(state)
+    storagePresent: isStoragePresent(state),
+    storageRoleName: rolesNames.storage,
+    routerRoleName: rolesNames.router
   }
 };